move CLI executable into bin/
